Keep each Promise regression scenario independently runnable
The generic applicability matrix had been appended as a second Test function in the mixed-receiver file. Move it unchanged into its own conventionally named file and complete the required scenario contract so focused runs and failures remain isolated.

Constraint: Go regression cases use one Test function per file

Rejected: Leave both matrices co-located | violates the package test layout and couples focused execution

Confidence: high

Scope-risk: narrow

Directive: Add new Test functions only in matching one-case files

Tested: git diff --check

Not-tested: Focused local tests remain deferred until three clean exact-head reviews

Related: #518
